fix:group and entry relation

+++ src/main/java/org/example/phonebookspring/service/GroupServiceImpl.java
@@ -1,9 +1,11 @@
 package org.example.phonebookspring.service;
 
+import org.example.phonebookspring.entity.Entry;
 import org.example.phonebookspring.entity.Group;
 import org.example.phonebookspring.repository.GroupRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Collection;
 
@@ -16,9 +18,16 @@ public class GroupServiceImpl implements  GroupService {
     public Group addGroup(Group group) {
         return groupRepository.save(group);
     }
-
+    @Transactional
     @Override
     public void deleteGroup(String GroupName) {
+        Group group = groupRepository.findByGroupName(GroupName);
+        if (group == null) {
+            return;
+        }
+        for(Entry entry: group.getEntries()){
+            group.removeFromEntry(entry);
+        }
         groupRepository.deleteGroupByGroupName(GroupName);
     }
